be-web/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.11
pydantic==2.9.0
pydantic-settings==2.5.0
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
python-multipart==0.0.9
alembic==1.13.0
slowapi==0.1.9
email-validator==2.1.0